// src/api/generated/model/character.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Dungeon Desk API
 * API documentation for Dungeon Desk application
 * OpenAPI spec version: 1.0.0
 */
import type { CharacterData } from './characterData';

export interface Character {
  id: number;
  name: string;
  race: string;
  class: string;
  /**
   * @minimum 1
   * @maximum 20
   */
  level: number;
  data: CharacterData;
  userId: number;
  createdAt: string;
  updatedAt: string;
}