// src/api/generated/model/characterCreateInput.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Dungeon Desk API
 * API documentation for Dungeon Desk application
 * OpenAPI spec version: 1.0.0
 */
import type { CharacterData } from './characterData';

export interface CharacterCreateInput {
  name: string;
  /**
   * @minimum 1
   * @maximum 20
   */
  level: number;
  data: CharacterData;
}